// types/task.ts
export enum TaskType {
  LAUNCH_BROWSER = 'LAUNCH_BROWSER',
  PAGE_TO_HTML = 'PAGE_TO_HTML',
  EXTRACT_TEXT_FROM_ELEMENT = 'EXTRACT_TEXT_FROM_ELEMENT',
  FILL_INPUT = 'FILL_INPUT',
  CLICK_ELEMENT = 'CLICK_ELEMENT',
  WAIT_FOR_ELEMENT = 'WAIT_FOR_ELEMENT',
  DELIVER_VIA_WEBHOOK = 'DELIVER_VIA_WEBHOOK',
  EXTRACT_DATA_WITH_AI = 'EXTRACT_DATA_WITH_AI',
  READ_PROPERTY_FROM_JSON = 'READ_PROPERTY_FROM_JSON',
  ADD_PROPERTY_TO_JSON = 'ADD_PROPERTY_TO_JSON',
  NAVIGATE_URL = 'NAVIGATE_URL',
  SCROLL_TO_ELEMENT = 'SCROLL_TO_ELEMENT',
  SCREENSHOT = 'SCREENSHOT',
  EVALUATE_JS = 'EVALUATE_JS',
  SET_VIEWPORT = 'SET_VIEWPORT',
  SET_USER_AGENT = 'SET_USER_AGENT',
  SET_COOKIES = 'SET_COOKIES',
  SET_LOCAL_STORAGE = 'SET_LOCAL_STORAGE',
  HOVER_ELEMENT = 'HOVER_ELEMENT',
  KEYBOARD_TYPE = 'KEYBOARD_TYPE',
  WAIT_FOR_NETWORK_IDLE = 'WAIT_FOR_NETWORK_IDLE',
  WAIT_FOR_NAVIGATION = 'WAIT_FOR_NAVIGATION',
  HTTP_REQUEST = 'HTTP_REQUEST',
  EXTRACT_ATTRIBUTES = 'EXTRACT_ATTRIBUTES',
  EXTRACT_LIST = 'EXTRACT_LIST',
  REGEX_EXTRACT = 'REGEX_EXTRACT',
  INFINITE_SCROLL = 'INFINITE_SCROLL',
  DELAY = 'DELAY',
  DISPLAY_DATA = 'DISPLAY_DATA',
  STORE_DATA = 'STORE_DATA',
  RETRIEVE_DATA = 'RETRIEVE_DATA',
}

export enum TaskParamType {
  STRING = 'STRING',
  BROWSER_INSTANCE = 'BROWSER_INSTANCE',
  SELECT = 'SELECT',
  CREDENTIAL = 'CREDENTIAL',
}

export interface TaskParam {
  name: string;
  type: TaskParamType;
  helperText?: string;
  required?: boolean;
  hideHandle?: boolean;
  [key: string]: any;
}